OpenCV_Assistant/requirements.txt
mss==9.0.1
numpy==1.25.2
opencv_contrib_python==4.7.0.72
Pillow==9.5.0
